Move rbac tests to security package
