+++ tests/integration/security/rbac_v1alpha1_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package rbac
+package security
 
 import (
 	"testing"
@@ -29,6 +29,11 @@ import (
 	"istio.io/istio/tests/common/jwt"
 	"istio.io/istio/tests/integration/security/util"
 	"istio.io/istio/tests/integration/security/util/connection"
+	"istio.io/istio/tests/integration/security/util/rbac_util"
+)
+
+const (
+	rbacClusterConfigTmpl = "testdata/rbac/clusterrbacconfig.yaml"
 )
 
 func TestV1_OptionalJWT(t *testing.T) {
@@ -47,7 +52,7 @@ func TestV1_OptionalJWT(t *testing.T) {
 				With(&b, util.EchoConfig("b", ns, false, nil, g, p)).
 				BuildOrFail(t)
 
-			cases := []TestCase{
+			cases := []rbac_util.TestCase{
 				{
 					Request: connection.Checker{
 						From: a,
@@ -80,12 +85,12 @@ func TestV1_OptionalJWT(t *testing.T) {
 			}
 			policies := tmpl.EvaluateAllOrFail(t, args,
 				file.AsStringOrFail(t, rbacClusterConfigTmpl),
-				file.AsStringOrFail(t, "testdata/v1-policy-optional-jwt.yaml.tmpl"))
+				file.AsStringOrFail(t, "testdata/rbac/v1-policy-optional-jwt.yaml.tmpl"))
 
 			g.ApplyConfigOrFail(t, ns, policies...)
 			defer g.DeleteConfigOrFail(t, ns, policies...)
 
-			RunRBACTest(t, cases)
+			rbac_util.RunRBACTest(t, cases)
 		})
 }
 
@@ -109,7 +114,7 @@ func TestV1_Group(t *testing.T) {
 				With(&c, util.EchoConfig("c", ns, false, nil, g, p)).
 				BuildOrFail(t)
 
-			cases := []TestCase{
+			cases := []rbac_util.TestCase{
 				{
 					Request: connection.Checker{
 						From: a,
@@ -169,12 +174,12 @@ func TestV1_Group(t *testing.T) {
 			}
 			policies := tmpl.EvaluateAllOrFail(t, args,
 				file.AsStringOrFail(t, rbacClusterConfigTmpl),
-				file.AsStringOrFail(t, "testdata/v1-policy-group.yaml.tmpl"))
+				file.AsStringOrFail(t, "testdata/rbac/v1-policy-group.yaml.tmpl"))
 
 			g.ApplyConfigOrFail(t, ns, policies...)
 			defer g.DeleteConfigOrFail(t, ns, policies...)
 
-			RunRBACTest(t, cases)
+			rbac_util.RunRBACTest(t, cases)
 		})
 }
 
@@ -194,7 +199,7 @@ func TestV1_GRPC(t *testing.T) {
 				With(&d, util.EchoConfig("d", ns, false, nil, g, p)).
 				BuildOrFail(t)
 
-			cases := []TestCase{
+			cases := []rbac_util.TestCase{
 				{
 					Request: connection.Checker{
 						From: b,
@@ -236,11 +241,11 @@ func TestV1_GRPC(t *testing.T) {
 
 			policies := tmpl.EvaluateAllOrFail(t, namespaceTmpl,
 				file.AsStringOrFail(t, rbacClusterConfigTmpl),
-				file.AsStringOrFail(t, "testdata/v1-policy-grpc.yaml.tmpl"))
+				file.AsStringOrFail(t, "testdata/rbac/v1-policy-grpc.yaml.tmpl"))
 			g.ApplyConfigOrFail(t, ns, policies...)
 			defer g.DeleteConfigOrFail(t, ns, policies...)
 
-			RunRBACTest(t, cases)
+			rbac_util.RunRBACTest(t, cases)
 		})
 }
 
@@ -280,8 +285,8 @@ func TestV1_Path(t *testing.T) {
 				}).
 				BuildOrFail(t)
 
-			newTestCase := func(path string, expectAllowed bool) TestCase {
-				return TestCase{
+			newTestCase := func(path string, expectAllowed bool) rbac_util.TestCase {
+				return rbac_util.TestCase{
 					Request: connection.Checker{
 						From: b,
 						Options: echo.CallOptions{
@@ -294,7 +299,7 @@ func TestV1_Path(t *testing.T) {
 					ExpectAllowed: expectAllowed,
 				}
 			}
-			cases := []TestCase{
+			cases := []rbac_util.TestCase{
 				newTestCase("/public", true),
 				newTestCase("/private", false),
 				newTestCase("/public/../private", false),
@@ -313,10 +318,10 @@ func TestV1_Path(t *testing.T) {
 			}
 			policies := tmpl.EvaluateAllOrFail(t, args,
 				file.AsStringOrFail(t, rbacClusterConfigTmpl),
-				file.AsStringOrFail(t, "testdata/v1-policy-path.yaml.tmpl"))
+				file.AsStringOrFail(t, "testdata/rbac/v1-policy-path.yaml.tmpl"))
 			g.ApplyConfigOrFail(t, ns, policies...)
 			defer g.DeleteConfigOrFail(t, ns, policies...)
 
-			RunRBACTest(t, cases)
+			rbac_util.RunRBACTest(t, cases)
 		})
 }
