+++ tests/integration/security/testdata/rbac/clusterrbacconfig.yaml
