+++ tests/integration/security/testdata/rbac/v1-policy-group.yaml.tmpl
