+++ tests/integration/security/testdata/rbac/v1-policy-grpc.yaml.tmpl
