+++ tests/integration/security/testdata/rbac/v1-policy-optional-jwt.yaml.tmpl
