+++ tests/integration/security/testdata/rbac/v1-policy-path.yaml.tmpl
