+++ tests/integration/security/testdata/rbac/v1beta1-override-v1alpha1.yaml.tmpl
