+++ tests/integration/security/testdata/rbac/v1beta1-workload-ns1.yaml.tmpl
