+++ tests/integration/security/testdata/rbac/v1beta1-workload-ns2.yaml.tmpl
