+++ tests/integration/security/util/rbac_util/util.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package rbac
+package rbac_util
 
 import (
 	"fmt"
